# src/main/java/cz/demo/usermanagement/controller/dto/SaveUserRequest.java
package cz.demo.usermanagement.controller.dto;

import jakarta.validation.constraints.NotBlank;
import lombok.Data;
import lombok.ToString;

@Data
public class SaveUserRequest {

    @NotBlank(message = "Name is mandatory")
    private String userName;

    private String password;

    @NotBlank(message = "First name is mandatory")
    private String firstName;

    @NotBlank(message = "Last name is mandatory")
    private String lastName;

    @ToString.Include(name = "password")
    public String getPasswordMasked() {
        return password != null ? "***" : "null";
    }
}